+++ SentinelOne/sentinelone_module/logs/connector.py
@@ -1,7 +1,6 @@
 import json
 from datetime import UTC, datetime, timedelta
 from functools import cached_property
-from logging import getLogger
 from threading import Event, Thread
 from time import sleep, time
 
@@ -16,13 +15,11 @@ from sekoia_automation.connector import Connector
 from sekoia_automation.storage import PersistentJSON
 
 from sentinelone_module.base import SentinelOneModule
+from sentinelone_module.logging import get_logger
 from sentinelone_module.logs.configuration import SentinelOneLogsConnectorConfiguration
 from sentinelone_module.logs.metrics import EVENTS_LAG, FORWARD_EVENTS_DURATION, OUTCOMING_EVENTS, INCOMING_MESSAGES
 
-# Declare prometheus metrics
-prom_namespace = "symphony_module_common"
-
-logger = getLogger()
+logger = get_logger()
 
 
 class SentinelOneLogsConsumer(Thread):
@@ -35,6 +32,7 @@ class SentinelOneLogsConsumer(Thread):
 
         self.context = PersistentJSON("context.json", connector._data_path)
         self.log = connector.log
+        self.log_exception = connector.log_exception
         self.configuration = connector.configuration
         self.connector = connector
         self.module = connector.module
@@ -133,28 +131,38 @@ class SentinelOneLogsConsumer(Thread):
         else:
             return latest_event_datetime
 
-    def pull_events(self):
+    def pull_events(self) -> list:
         raise NotImplementedError
 
     def next_batch(self):
         # save the starting time
         batch_start_time = time()
 
-        # get the batch
-        self.pull_events()
+        try:
+            # get the batch
+            events_id = self.pull_events()
 
-        # get the ending time and compute the duration to fetch the events
-        batch_end_time = time()
-        batch_duration = int(batch_end_time - batch_start_time)
-        logger.debug(f"Fetched and forwarded events in {batch_duration} seconds")
-        FORWARD_EVENTS_DURATION.labels(intake_key=self.configuration.intake_key, datasource="sentinelone").observe(
-            batch_duration
-        )
+            # get the ending time and compute the duration to fetch the events
+            batch_end_time = time()
+            batch_duration = int(batch_end_time - batch_start_time)
+            logger.debug(f"Fetched and forwarded events", duration=batch_duration, nb_events=len(events_id))
+            FORWARD_EVENTS_DURATION.labels(intake_key=self.configuration.intake_key, datasource="sentinelone").observe(
+                batch_duration
+            )
+
+            # log the number of forwarded events
+            log_message = "No events to forward"
+            if len(events_id) > 0:
+                log_message = f"Fetched and forwarded {len(events_id)} events"
+
+            self.log(message=log_message, level="info")
+        except Exception as ex:
+            self.log_exception(ex, message="Failed to forward events")
 
         # compute the remaining sleeping time. If greater than 0, sleep
         delta_sleep = self.configuration.frequency - batch_duration
         if delta_sleep > 0:
-            logger.debug(f"Next batch in the future. Waiting {delta_sleep} seconds")
+            logger.debug(f"Next batch in the future. Pause the connector", pause=delta_sleep)
             sleep(delta_sleep)
 
     def run(self):
@@ -176,7 +184,7 @@ class SentinelOneLogsConsumer(Thread):
 
 
 class SentinelOneActivityLogsConsumer(SentinelOneLogsConsumer):
-    def pull_events(self):
+    def pull_events(self) -> list:
         """Fetches activities from SentinelOne"""
         # Set  filters
         query_filter = ActivitiesFilter()
@@ -184,17 +192,19 @@ class SentinelOneActivityLogsConsumer(SentinelOneLogsConsumer):
         query_filter.apply(key="sortBy", val="createdAt")
         query_filter.apply(key="sortOrder", val="asc")
 
+        events_id = []
         while self.running:
             # Fetch activities
             activities = self.management_client.activities.get(query_filter)
-            logger.debug("activities: received %d events" % len(activities.data))
+            nb_activities = len(activities.data)
+            logger.debug("Collected activities", nb=nb_activities)
 
             INCOMING_MESSAGES.labels(intake_key=self.configuration.intake_key, datasource="sentinelone").inc(
-                len(activities.data)
+                nb_activities
             )
 
             # Push events
-            self.connector.push_events_to_intakes(self._serialize_events(activities.data))
+            events_id.extend(self.connector.push_events_to_intakes(self._serialize_events(activities.data)))
 
             # Update context with latest event date
             latest_event_timestamp = self._get_latest_event_timestamp(activities.data)
@@ -203,10 +213,10 @@ class SentinelOneActivityLogsConsumer(SentinelOneLogsConsumer):
 
             # Send Prometheus metrics
             OUTCOMING_EVENTS.labels(intake_key=self.configuration.intake_key, datasource="sentinelone").inc(
-                len(activities.data)
+                nb_activities
             )
 
-            if len(activities.data) > 0:
+            if nb_activities > 0:
                 EVENTS_LAG.labels(intake_key=self.configuration.intake_key, type="activities").set(
                     (datetime.now(UTC) - latest_event_timestamp).total_seconds()
                 )
@@ -219,6 +229,8 @@ class SentinelOneActivityLogsConsumer(SentinelOneLogsConsumer):
 
             query_filter.apply(key="cursor", val=activities.pagination["nextCursor"])
 
+        return events_id
+
 
 class SentinelOneThreatLogsConsumer(SentinelOneLogsConsumer):
     def pull_events(self):
@@ -228,13 +240,15 @@ class SentinelOneThreatLogsConsumer(SentinelOneLogsConsumer):
         query_filter.apply(key="sortBy", val="createdAt")
         query_filter.apply(key="sortOrder", val="asc")
 
+        events_id = []
         while self.running:
             # Fetch threats
             threats = self.management_client.client.get(endpoint="threats", params=query_filter.filters)
-            logger.debug("threats: received %d events" % len(threats.data))
+            nb_threats = len(threats.data)
+            logger.debug("Collected nb_threats", nb=nb_threats)
 
             # Push events
-            self.connector.push_events_to_intakes(self._serialize_events(threats.data))
+            events_id.extend(self.connector.push_events_to_intakes(self._serialize_events(threats.data)))
 
             # Update context with the latest event date
             latest_event_timestamp = self._get_latest_event_timestamp(threats.data)
@@ -242,11 +256,9 @@ class SentinelOneThreatLogsConsumer(SentinelOneLogsConsumer):
                 cache["last_event_date"] = latest_event_timestamp.isoformat()
 
             # Send Prometheus metrics
-            OUTCOMING_EVENTS.labels(intake_key=self.configuration.intake_key, datasource="sentinelone").inc(
-                len(threats.data)
-            )
+            OUTCOMING_EVENTS.labels(intake_key=self.configuration.intake_key, datasource="sentinelone").inc(nb_threats)
 
-            if len(threats.data) > 0:
+            if nb_threats > 0:
                 EVENTS_LAG.labels(intake_key=self.configuration.intake_key, type="threats").set(
                     (datetime.now(UTC) - latest_event_timestamp).total_seconds()
                 )
@@ -258,6 +270,8 @@ class SentinelOneThreatLogsConsumer(SentinelOneLogsConsumer):
                 break
 
             query_filter.apply(key="cursor", val=threats.pagination["nextCursor"])
+
+        return events_id
 
 
 CONSUMER_TYPES = {"activity": SentinelOneActivityLogsConsumer, "threat": SentinelOneThreatLogsConsumer}